% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps.to.amino.R
\name{snps.to.amino}
\alias{snps.to.amino}
\title{SNPs to Amino Acids}
\usage{
snps.to.amino(snp_db, ref_seq, cores = parallelly::availableCores())
}
\arguments{
\item{snp_db}{A data frame with a column named "SNP"}

\item{ref_seq}{A path to a reference sequence (.gb) format.}

\item{cores}{Option to specify number of cores to use.}
}
\value{
a data frame with SNP columns and an AminoAcid column
}
\description{
This function takes a data frame from fasta.to.snps or any data frame with a column named "SNP" and SNPS separated as "Reference_Location_SNP"
}
